Update fixtures to current user fixtures and remove skip from tests
