+++ playwright/e2e/global-admin/detected-vulnerabilities-widget.spec.ts
@@ -3,13 +3,13 @@ import { expect } from '@playwright/test';
 import { ROUTES } from '../../../frontend/src/constants/routes';
 
 test.describe('Known Exploited and Other Detected Vulnerabilities Widget', () => {
-  test.beforeEach(async ({ page: pageAsGlobalAdmin }) => {
+  test.beforeEach(async ({ pageAsGlobalAdmin }) => {
     // Navigate to the dashboard page before each test
     await pageAsGlobalAdmin.goto(ROUTES.VSDASHBOARD);
   });
 
-  test.skip('should display main widget heading and tooltip', async ({
-    page: pageAsGlobalAdmin
+  test('should display main widget heading and tooltip', async ({
+    pageAsGlobalAdmin
   }) => {
     // Check that the main widget heading is visible
     const heading = pageAsGlobalAdmin.getByRole('heading', {
@@ -24,8 +24,8 @@ test.describe('Known Exploited and Other Detected Vulnerabilities Widget', () =>
     await tooltipButton.click();
   });
 
-  test.skip('should navigate to details page when clicking "View Details"', async ({
-    page: pageAsGlobalAdmin
+  test('should navigate to details page when clicking "View Details"', async ({
+    pageAsGlobalAdmin
   }) => {
     // Verify "View Details" link is visible and navigates correctly
     const detailsLink = pageAsGlobalAdmin
@@ -38,8 +38,8 @@ test.describe('Known Exploited and Other Detected Vulnerabilities Widget', () =>
 
   //========= Severity by Prominenece ==========
 
-  test.skip('should toggle severity by prominence graph data using KEV, Distinct, All buttons', async ({
-    page: pageAsGlobalAdmin
+  test('should toggle severity by prominence graph data using KEV, Distinct, All buttons', async ({
+    pageAsGlobalAdmin
   }) => {
     // Scope radios to the "Severity by Prominence" section
     const heading = pageAsGlobalAdmin.getByRole('heading', {
@@ -63,8 +63,8 @@ test.describe('Known Exploited and Other Detected Vulnerabilities Widget', () =>
     await expect(allBtn).toHaveAttribute('aria-checked', 'true');
   });
 
-  test.skip('should render severity bars with correct labels', async ({
-    page: pageAsGlobalAdmin
+  test('should render severity bars with correct labels', async ({
+    pageAsGlobalAdmin
   }) => {
     // Locate all bars on the graph with aria-labels
     const bars = pageAsGlobalAdmin.locator(
@@ -79,8 +79,8 @@ test.describe('Known Exploited and Other Detected Vulnerabilities Widget', () =>
     await expect(bars.nth(3)).toHaveAttribute('aria-label', /low/i);
   });
 
-  test.skip(' severity chart bar colors are correct for KEV, Distinct, and All', async ({
-    page: pageAsGlobalAdmin
+  test(' severity chart bar colors are correct for KEV, Distinct, and All', async ({
+    pageAsGlobalAdmin
   }) => {
     const bars = pageAsGlobalAdmin.locator('svg .MuiBarElement-root');
     //await expect(bars).toHaveCount(4);
@@ -106,8 +106,8 @@ test.describe('Known Exploited and Other Detected Vulnerabilities Widget', () =>
 
   // ====== Top Vulnerability by Occurrence =======
 
-  test.skip('should toggle  occurrence table data using KEV and All buttons', async ({
-    page: pageAsGlobalAdmin
+  test('should toggle  occurrence table data using KEV and All buttons', async ({
+    pageAsGlobalAdmin
   }) => {
     // Scope radios to the "Top Vulnerabilities by Occurrence" section
     const heading = pageAsGlobalAdmin.getByRole('heading', {
@@ -127,8 +127,8 @@ test.describe('Known Exploited and Other Detected Vulnerabilities Widget', () =>
     await expect(allBtn).toHaveAttribute('aria-checked', 'true');
   });
 
-  test.skip('should display vulnerability occurrunce table with correct columns and data', async ({
-    page: pageAsGlobalAdmin
+  test('should display vulnerability occurrunce table with correct columns and data', async ({
+    pageAsGlobalAdmin
   }) => {
     // Assert the table headers are visible
     const table = pageAsGlobalAdmin.locator('table');
@@ -153,8 +153,8 @@ test.describe('Known Exploited and Other Detected Vulnerabilities Widget', () =>
     await firstCell.click();
   });
 
-  test.skip('hover tooltips show text for all info icons', async ({
-    page: pageAsGlobalAdmin
+  test('hover tooltips show text for all info icons', async ({
+    pageAsGlobalAdmin
   }) => {
     // Small helper to reduce flakiness with MUI’s show/leave delays
     const resetHover = async () => {
